feat: separated sharding operations into its own file module
