+++ swarm_nl/src/core/replication.rs
@@ -1,3 +1,6 @@
+// Copyright 2024 Algorealm, Inc.
+// Apache 2.0 License
+
 //! Module that contains important data structures to manage `Replication` operations on the
 //! network.
